Extract AuthorizeResult from install-provider.ts
